extract quiz choice button
